auth_service/requirements.txt
fastapi
uvicorn
pymongo
requests
python-multipart
prometheus_fastapi_instrumentator
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http